PDF_OCR/requirements_easyocr.txt
easyocr>=1.7.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
Pillow>=10.0.0
numpy>=1.21.0



